extract struct
