+++ libs/targets/gesha-rust-types/src/definition.rs
@@ -1,6 +1,6 @@
 use crate::{
-    DataType, Definitions, EnumVariant, ErrorDef, Imports, MediaTypeDef, ModuleName,
-    RequestBodyDef, StructField, TypeHeader,
+    DataType, Definitions, EnumDef, ErrorDef, Imports, MediaTypeDef, ModuleName, RequestBodyDef,
+    StructField, TypeHeader,
 };
 
 #[derive(Clone, Debug, PartialEq)]
@@ -87,29 +87,6 @@ impl NewTypeDef {
 impl From<NewTypeDef> for Definition {
     fn from(x: NewTypeDef) -> Self {
         Self::NewTypeDef(x)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct EnumDef {
-    pub header: TypeHeader,
-    pub variants: Vec<EnumVariant>,
-    _hide_default_constructor: bool,
-}
-
-impl EnumDef {
-    pub fn new(header: TypeHeader, variants: Vec<EnumVariant>) -> Self {
-        Self {
-            header,
-            variants,
-            _hide_default_constructor: true,
-        }
-    }
-}
-
-impl From<EnumDef> for Definition {
-    fn from(this: EnumDef) -> Self {
-        Self::EnumDef(this)
     }
 }
