+++ libs/targets/gesha-rust-types/src/enum_def.rs
@@ -1,0 +1,24 @@
+use crate::{Definition, EnumVariant, TypeHeader};
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct EnumDef {
+    pub header: TypeHeader,
+    pub variants: Vec<EnumVariant>,
+    _hide_default_constructor: bool,
+}
+
+impl EnumDef {
+    pub fn new(header: TypeHeader, variants: Vec<EnumVariant>) -> Self {
+        Self {
+            header,
+            variants,
+            _hide_default_constructor: true,
+        }
+    }
+}
+
+impl From<EnumDef> for Definition {
+    fn from(this: EnumDef) -> Self {
+        Self::EnumDef(this)
+    }
+}
